Changed naming convention from WebSocketCarbonMessage to WebSocketMessage. Added (.)dot to the ends of comments. Added new doc comments  and expanded some comments to make the implementation clear.
